Added support for reading selected FITS extensions for all input images/maps (e.g. image.fits[3] a la CFITSIO). Fixed some (well deserved) gcc -Wall warnings. Added new ATLAS handling features to configure (TBC). Pushed version number to 2.6.3.

+++ branches/sexfigi/src/profit.h
@@ -9,7 +9,7 @@
 *
 *	Contents:	Include file for profit.c.
 *
-*	Last modify:	08/08/2007
+*	Last modify:	08/10/2007
 *
 *%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
 */
@@ -113,7 +113,7 @@ typedef struct
   struct psf	*psf;		/* PSF */
   double	*psfdft;	/* Compressed Fourier Transform of the PSF */
   double	*modpix;	/* Full res. pixmap of the complete model */
-  double	*pmodpix;	/* Full res. pixmap of the partial model */
+  float		*pmodpix;	/* Full res. pixmap of the partial model */
   int		modnaxisn[2];	/* Dimensions along each axis */
   PIXTYPE	*lmodpix;	/* Low resolution pixmap of the model */
   PIXTYPE	*objpix;	/* Copy of object pixmap */
